Implemented Console GUI
Added console GUI, board, square, and chessgame classes. After reading
in the actions the program now performs it on the board, unless
something prevents it where it returns a message, then displays the new
state of the board after the action is performed. Main method is now in
the ChessGame class.